Extract server-related components into server app
This refactoring separates server functionalities (Discovery, Registries, Repositories) from the core SDK to improve modularity and maintainability.

Changes:
- Server is now a separate app
- Added an initial pyproject.toml for the server app
- Moved all server-related classes and functions from sdk to server
- Consolidated descriptor.py, submodel_descriptor.py, and aas_descriptor.py into a single server_model.py file
