+++ server/pyproject.toml
@@ -17,12 +17,12 @@ build-backend = "setuptools.build_meta"
 # print(f"Project version: {version}")
 # ```
 root = ".."  # Defines the path to the root of the repository
-version_file = "basyx/version.py"
+version_file = "app/version.py"
 
 [project]
-    name = "basyx-python-sdk"
+name = "basyx-python-server"
 dynamic = ["version"]
-description = "The Eclipse BaSyx Python SDK, an implementation of the Asset Administration Shell for Industry 4.0 systems"
+description = "The Eclipse BaSyx Python Server, an implementation of the BaSyx AAS Server" #FIXME
 authors = [
     { name = "The Eclipse BaSyx Authors", email = "admins@iat.rwth-aachen.de" }
 ]
@@ -36,9 +36,9 @@ classifiers = [
 ]
 requires-python = ">=3.9"
 dependencies = [
-    "lxml>=4.2,<5",
-    "python-dateutil>=2.8,<3",
-    "pyecma376-2>=1.0.1"
+    "basyx-python-sdk", #TODO: Think about the version
+    "urllib3>=1.26,<3",
+    "Werkzeug>=3.0.3,<4",
 ]
 
 [project.optional-dependencies]
@@ -51,7 +51,6 @@ dev = [
     "jsonschema~=4.7",
     "hypothesis~=6.13",
     "types-python-dateutil",
-    "lxml-stubs~=0.5.1",
 ]
 
 [project.urls]
@@ -61,5 +60,4 @@ dev = [
 packages = { find = { include = ["basyx*"], exclude = ["test*"] } }
 
 [tool.setuptools.package-data]
-basyx = ["py.typed"]
-"basyx.aas.examples.data" = ["TestFile.pdf"]
+app = ["py.typed"]
